test/web-platform-tests/to-upstream/html/webappapis/scripting/processing-model-2/window-onerror-event-throw-null.html

<!DOCTYPE html>
<meta charset="UTF-8">
<title>window.onerror: throwing null in event listeners</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
  "use strict";

  setup({ allow_uncaught_exception: true });

  test(() => {
    let errorEvents = 0;
    window.addEventListener("error", event => {
      assert_equals(event.message, null);
      assert_equals(event.error, null);
      ++errorEvents;
    });

    const element = document.createElement("div");

    element.addEventListener("click", () => {
      throw null; // eslint-disable-line no-throw-literal
    });

    element.dispatchEvent(new Event("click"));

    assert_equals(errorEvents, 1);
  }, "Throwing null in event listener");
</script>
